Split up Model_Simulation and Model_Setup
